move to cmd/ and pkg/healthmonitor, pkg/loadbalancer
