+++ .gitignore
@@ -2,3 +2,4 @@ config.yaml
 proxy
 vendor
 tcp-mux-proxy
+!/cmd/tcp-mux-proxy/main.go
