+++ pkg/loadbalancer/pow_of_two_lb.go
@@ -1,4 +1,4 @@
-package main
+package loadbalancer
 
 import (
 	"errors"
